fix(kml): address parser review feedback
