+++ modules/kml/src/index.ts
@@ -9,8 +9,8 @@ export {GPXLoader} from './gpx-loader';
 export type {KMLLoaderOptions} from './kml-loader';
 export {KMLLoader} from './kml-loader';
 
-export type {KMZLoaderOptions} from './kmz-loader';
-export {KMZLoader} from './kmz-loader';
+export type {KMZLoaderOptions} from './kmz-loader-types';
+export {KMZLoader} from './kmz-loader-types';
 
 export type {KMZArchive} from './kmz-archive';
 export {openKMZArchive, resolveKMZResourcePath} from './kmz-archive';
