+++ tsconfig.json
@@ -202,6 +202,9 @@
       "@loaders.gl/kml": [
         "modules/kml/src"
       ],
+      "@loaders.gl/kml/*": [
+        "modules/kml/src/*"
+      ],
       "@loaders.gl/kml/test": [
         "modules/kml/test"
       ],
